break header and footer into components
