+++ app/root.tsx
@@ -1,7 +1,6 @@
-import { LoaderFunction, useSubmit } from "remix";
+import React from "react";
 import {
   json,
-  Link,
   Links,
   LiveReload,
   Meta,
@@ -11,15 +10,13 @@ import {
   useCatch,
   useLoaderData,
 } from "remix";
-import type { MetaFunction, LinksFunction } from "remix";
-import { ChevronDownIcon } from "@heroicons/react/solid";
-import { Menu } from "@headlessui/react";
-
-import tailwindStyles from "./tailwind.css";
-import React from "react";
+import type { LoaderFunction, MetaFunction, LinksFunction } from "remix";
 import type { User } from "@prisma/client";
-import { getUser } from "./utils/session.server";
-import clsx from "clsx";
+
+import { getUser } from "~/utils/session.server";
+import { Header } from "~/components/Header";
+import { Footer } from "~/components/Footer";
+import tailwindStyles from "~/tailwind.css";
 
 export const links: LinksFunction = () => [
   { rel: "stylesheet", href: tailwindStyles },
@@ -47,13 +44,6 @@ function Document({
   error?: boolean;
   children: React.ReactNode;
 }) {
-  const submit = useSubmit();
-
-  function handleLogoutClick() {
-    submit(new FormData(), { action: "/logout", method: "post" });
-  }
-
-  // TODO break header and footer into their own components instead of polluting here
   return (
     <html lang="en" className="h-full">
       <head>
@@ -64,99 +54,10 @@ function Document({
       </head>
       <body className="bg-dark-500 p-4 h-full">
         <div className="mx-auto px-4 min-h-[91.5vh] max-w-7xl">
-          <header className="border-b border-light p-2 flex justify-between max-w-[36rem] mx-auto">
-            <Link to="/" className="hover:underline flex items-center gap-2">
-              <img src="/favicon.svg" alt="stickertrade logo" className="h-4" />
-              <h1>stickertrade</h1>
-            </Link>
-            <div className="flex flex-col gap-4 items-center">
-              {!error &&
-                (user ? (
-                  <div className="flex items-center gap-4">
-                    <Menu
-                      as="div"
-                      className="relative h-0 inline-block text-left"
-                    >
-                      <Menu.Button className="inline-flex items-center justify-center w-full">
-                        <div className="flex items-center gap-3">
-                          <img
-                            className="w-[1.6em] rounded-full"
-                            src={
-                              user.avatarUrl ?? "/images/default-avatar.webp"
-                            }
-                            alt={user.username}
-                          />
-                          <p>{user.username}</p>
-                        </div>
-                        <ChevronDownIcon
-                          className="w-5 h-5 ml-1 mt-0.5"
-                          aria-hidden="true"
-                        />
-                      </Menu.Button>
-                      <Menu.Items className="absolute -mt-1 right-0 origin-top-right bg-light-500 divide-y divide-dark-100 rounded-sm focus:outline-none">
-                        <div className="px-1 py-1 ">
-                          <Menu.Item>
-                            {({ active }) => (
-                              <Link to={`/profile/${user.username}`}>
-                                <button
-                                  className={clsx(
-                                    { "bg-primary-400": active },
-                                    "text-dark-500 group flex rounded-sm items-center w-full px-2 py-1.5 text-sm"
-                                  )}
-                                >
-                                  profile
-                                </button>
-                              </Link>
-                            )}
-                          </Menu.Item>
-                        </div>
-                        <div className="px-1 py-1 ">
-                          <Menu.Item>
-                            {({ active }) => (
-                              <button
-                                type="submit"
-                                onClick={handleLogoutClick}
-                                className={clsx(
-                                  { "bg-primary-400": active },
-                                  "text-dark-500 group flex rounded-sm items-center w-full px-2 py-1.5 text-sm"
-                                )}
-                              >
-                                logout
-                              </button>
-                            )}
-                          </Menu.Item>
-                        </div>
-                      </Menu.Items>
-                    </Menu>
-                  </div>
-                ) : (
-                  <Link to="/login" className="hover:underline">
-                    <h1>login</h1>
-                  </Link>
-                ))}
-            </div>
-          </header>
+          <Header user={user} error={error} />
           <div className="pt-5 pb-8">{children}</div>
         </div>
-        <footer className="mb-2 border-t mx-auto max-w-[36rem] border-t-light-500">
-          <div className="flex flex-col sm:flex-row gap-2 sm:gap-4 py-3 sm:py-2 p-2 justify-between">
-            <Link to="/" className="hover:underline flex items-center gap-2">
-              <img src="/favicon.svg" alt="stickertrade logo" className="h-4" />
-              <h1>stickertrade</h1>
-            </Link>
-            <div className="flex flex-col sm:flex-row gap-2 sm:gap-4">
-              <Link to="/roadmap" className="hover:underline">
-                <h1>roadmap</h1>
-              </Link>
-              <Link to="/brand" className="hover:underline">
-                <h1>brand</h1>
-              </Link>
-              <Link to="/dev-logs" className="hover:underline">
-                <h1>dev logs</h1>
-              </Link>
-            </div>
-          </div>
-        </footer>
+        <Footer />
         <ScrollRestoration />
         <Scripts />
         <LiveReload />
